requirements.txt
fastapi
uvicorn
pillow
opencv-python
rembg
numpy
realesrgan
python-multipart